move deleteVersionedMeasures to measure service
